pyproject.toml
[tool.poetry]
name = "Mugen"
version = "0.1.0"
description = ""
authors = ["FallenDeity <61227305+FallenDeity@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pygame = "^2.5.2"
moderngl = "^5.9.0"
pyglm = "^2.7.1"
numba = "^0.58.1"
python-dotenv = "^1.0.0"
opensimplex = "^0.4.5"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
isort = "^5.13.2"
ruff = "^0.1.9"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
nptyping = "^2.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 120
ignore = ["F722"]

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
src_paths = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
